requirements.txt
fastapi
uvicorn[standard]
onnxruntime
rembg
httpx
pillow
numpy